requirements.txt
psutil>=5.9.0
mss>=9.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
sounddevice>=0.4.8
numpy>=1.24.0
pyautogui>=0.9.53
pynput>=1.7.6
